Added compatibility with Logging based loggers
This commits support for loggers based on the Logging library, such as Fern and Lib4Rs. In order to set up a logger, you set it up in the main method in your expr library. This allows us to get pretty formatted output, and filtering of output based on output level, and chaining to both stdout and file output easily.